added botsettingsbase.cs
