+++ solutions/Virtual-Assistant/src/csharp/experimental/skills/restaurantbooking/Services/BotServices.cs
@@ -4,7 +4,7 @@ using Microsoft.Bot.Builder.AI.QnA;
 using Microsoft.Bot.Builder.Solutions;
 using Microsoft.Bot.Builder.Solutions.Telemetry;
 
-namespace RestaurantSkill.Services
+namespace RestaurantBooking.Services
 {
     public class BotServices
     {
