+++ solutions/Virtual-Assistant/src/csharp/skills/automotiveskill/automotiveskill/Models/ScoredMatchableBagOfTokens.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using AutomotiveSkill.Utilities;
-
 namespace AutomotiveSkill.Models
 {
     public class ScoredMatchableBagOfTokens
