+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/calendarskill/Services/BotSettings.cs
@@ -1,52 +1,8 @@
-﻿using System.Collections.Generic;
-using Microsoft.ApplicationInsights.Extensibility;
-using Microsoft.Bot.Builder.Azure;
-using Microsoft.Bot.Builder.Solutions.Shared.Authentication;
-using Microsoft.Bot.Configuration;
+﻿using Microsoft.Bot.Builder.Solutions.Shared;
 
 namespace CalendarSkill.Services
 {
-    public class BotSettings
+    public class BotSettings : BotSettingsBase
     {
-        public string MicrosoftAppId { get; set; }
-
-        public string MicrosoftAppPassword { get; set; }
-
-        public string DefaultLocale { get; set; }
-
-        public List<OAuthConnection> OAuthConnections { get; set; }
-
-        public CosmosDbStorageOptions CosmosDb { get; set; }
-
-        public TelemetryConfiguration AppInsights { get; set; }
-
-        public BlobStorageConfiguration BlobStorage { get; set; }
-
-        public ContentModeratorConfiguration ContentModerator { get; set; }
-
-        public Dictionary<string, CognitiveModelConfiguration> CognitiveModels { get; set; }
-
-        public Dictionary<string, string> Properties { get; set; }
-
-        public class BlobStorageConfiguration
-        {
-            public string ConnectionString { get; set; }
-
-            public string Container { get; set; }
-        }
-
-        public class ContentModeratorConfiguration
-        {
-            public string Key { get; set; }
-        }
-
-        public class CognitiveModelConfiguration
-        {
-            public DispatchService DispatchModel { get; set; }
-
-            public List<LuisService> LanguageModels { get; set; }
-
-            public List<QnAMakerService> Knowledgebases { get; set; }
-        }
     }
 }
