+++ templates/Virtual-Assistant-Template/VirtualAssistantTemplate/VirtualAssistantTemplate.Tests/Mocks/MockQnAMaker.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.AI.QnA;
-using Microsoft.Bot.Builder.Solutions.Telemetry;
+using Microsoft.Bot.Builder.Solutions.Shared.Telemetry;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
